Tuning

+++ config.go
@@ -3,5 +3,5 @@ package main
 type Config struct {
 	ffmpeg    string
 	ffprobe   string
-	chunkSize float64
+	chunkSize int
 }
